requirements.txt
python-dotenv
google-generativeai
IMDbPY